Break controller into base and type controllers
Also moves the views into the correct subfolders
and fixes routes etc.
