+++ app/views/candidate_interface/decoupled_references/base/start.html.erb
@@ -17,6 +17,6 @@
 
     <p class="govuk-body">You can add as many referees as you like to increase the chances of getting 2 references quickly.</p>
 
-    <%= govuk_link_to 'Continue', candidate_interface_decoupled_references_type_path, class: 'govuk-button' %>
+    <%= govuk_link_to 'Continue', candidate_interface_decoupled_references_new_type_path, class: 'govuk-button' %>
   </div>
 </div>
